plugins/simple_fastapi_dashboard/0.0.1/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "simple_fastapi_dashboard"
version = "0.0.1"
authors = [
    {name = 'Vladimir Molodtsov', email = 'vmol@mail.ru'},
]
keywords = ["peacepie"]
dependencies = [
    'fastapi',
    'jinja2',
    'pyjwt',
    'python-multipart',
    'pyzmq',
    'uvicorn[standard]',
    'pydantic',
]

[tool.setuptools.packages.find]
where = ["src"]
[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["templates/*", "templates/**/*"]